🦄 refactor(Refactor upload field component and add avatar editor dialog):
