+++ src/component/upload-field/index.ts
@@ -1,0 +1,3 @@
+export { default as UploadField } from './upload-field'
+export { default as AvatarEditorDialog } from './avatar-editor-dialog'
+export * from './style'
