--- src/component/uploadfield/index.ts
@@ -1,1 +1,0 @@
-export { default as UploadField } from './uploadfield'
